+++ src/utils/index.ts
@@ -6,3 +6,4 @@
  export const MAX_ERROR_COUNT = 2;
  export * from './guidHelper';
  export * from './luisAppSetup';
+ export * from './validateCanadianPhoneNumber';
